@@ -1,26 +1,25 @@
 name: fitnessapp
-description: "A new Flutter project."
+description: "A fitness application with an interactive and modern design."
 
 publish_to: 'none'
 
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.19.0 <4.0.0"
+  sdk: ^3.6.1
 
 dependencies:
   flutter:
     sdk: flutter
   cupertino_icons: ^1.0.8
-  # Ajustamos leak_tracker a una versión compatible con vm_service.
-  leak_tracker: ^10.0.7
-  # Ajustamos vm_service para que sea compatible con leak_tracker.
-  vm_service: ^14.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
+  assets:
+    - assets/images/
+    
