@@ -7,35 +7,57 @@ class MainScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Pantalla Principal"),
+        backgroundColor: Colors.teal,
+        title: const Text('Pantalla Principal'),
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
+            Image.asset('assets/images/calendario_app.jpeg', width: 150, height: 150),
+            const SizedBox(height: 20),
             ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.teal,
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              ),
               onPressed: () {
                 Navigator.pushNamed(context, '/createRoutine');
               },
-              child: const Text("Crear Rutina"),
+              child: const Text('Crear Rutina'),
             ),
+            const SizedBox(height: 10),
             ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.teal,
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              ),
               onPressed: () {
                 Navigator.pushNamed(context, '/recommended');
               },
-              child: const Text("Rutinas Recomendadas"),
+              child: const Text('Rutinas Recomendadas'),
             ),
+            const SizedBox(height: 10),
             ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.teal,
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              ),
               onPressed: () {
                 Navigator.pushNamed(context, '/progress');
               },
-              child: const Text("Ver Progreso"),
+              child: const Text('Ver Progreso'),
             ),
+            const SizedBox(height: 10),
             ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.teal,
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              ),
               onPressed: () {
                 Navigator.pushNamed(context, '/settings');
               },
-              child: const Text("Configuración"),
+              child: const Text('Configuración'),
             ),
           ],
         ),
