@@ -7,12 +7,17 @@ class SettingsScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Configuración"),
+        backgroundColor: Colors.teal,
+        title: const Text('Configuración'),
       ),
-      body: const Center(
-        child: Text(
-          "Pantalla de Configuración",
-          style: TextStyle(fontSize: 24),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Image.asset('assets/images/perfil_app.avif', width: 150, height: 150),
+            const SizedBox(height: 20),
+            const Text('Ajusta tus configuraciones aquí', style: TextStyle(fontSize: 18)),
+          ],
         ),
       ),
     );
