@@ -7,21 +7,30 @@ class WelcomeScreen extends StatelessWidget {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: const Text("Bienvenido"),
+        backgroundColor: Colors.teal,
+        title: const Text('Bienvenido', style: TextStyle(fontSize: 24)),
       ),
       body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
+            Image.asset('assets/images/perfil_app.avif', width: 150, height: 150),
+            const SizedBox(height: 20),
             const Text(
-              "Pantalla de Bienvenida",
-              style: TextStyle(fontSize: 24),
+              'Pantalla de Bienvenida',
+              style: TextStyle(fontSize: 22, fontWeight: FontWeight.bold, color: Colors.teal),
             ),
+            const SizedBox(height: 30),
             ElevatedButton(
+              style: ElevatedButton.styleFrom(
+                backgroundColor: Colors.teal,
+                padding: const EdgeInsets.symmetric(horizontal: 20, vertical: 10),
+                shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(10)),
+              ),
               onPressed: () {
                 Navigator.pushNamed(context, '/main');
               },
-              child: const Text("Ir a la Pantalla Principal"),
+              child: const Text('Ir a la Pantalla Principal', style: TextStyle(fontSize: 18)),
             ),
           ],
         ),
